Module chrome - adds new module chrome to replicate core Bootstrap fucntions
This is available in J3 only.

-t3tabs
Takes the title of any module published to the same position with the
t3tabs chrome and turns them into bootstrap tabs.

-t3slider
Turns the module into a bootstrap collapsable module where the title
triggers the open / collapse of the module body.

-t3modal
The title of the module is rendered as a btn class and when clicked
opens the contents of the module in a modal.

-popover
The title of the module is displayed in a btn format and when clicked
the contents of the moduel are displayed in the Bootstrap popover
format.

In addition to this you can use left, right, top, bottom to in the
moduel class to position the popover.