Adds a check for layout-custom.css in template admin folder.
In some of our themes the layout panel in the admin isnt displaying
properly or there are some cases where we want to hide certain elements.

This pull request adds a check for layout-custom.css in the template
admin folder which will aloow us to add css to control the display of
the layout in the admin.