Display menus per viewlevel
This fix displays the menus associated with the view level selected as
per J2.5/J3 view level model.

A further improvement might be to use the usergroup as the dropdown and
display the menus as seen by a user in that group, however this would be
an enhancement rather than a fix,  and is outside the scope of this fix.