Off-Canvas Gltch
when a flyout on the mainmenu is opened, there is an overflow of the hidden div animation which reveals itself on top of the mainmenu. In other words, when the submenu flys in from the top, the text rides on top of the parent menu container. It looks just a little sloppy